Split ProgramRunner immediate command handling (#4111)
