+++ changelog.d/unreleased/4111.internal.md
@@ -1,0 +1,16 @@
+---
+category: internal
+issues:
+  - 4111
+affected:
+  - src/CodeIndex/Cli/ProgramRunner.Dispatch.cs
+  - src/CodeIndex/Cli/ProgramRunner.ImmediateCommands.cs
+---
+
+## English
+
+- **ProgramRunner immediate command handling now has its own partial file (#4111)** — startup-only commands such as help, version, license, completions, doctor, and easter egg handling are separated from the regular CLI dispatcher to keep dispatch boundaries easier to test and review.
+
+## 日本語
+
+- **ProgramRunner の即時コマンド処理を専用 partial ファイルへ分離しました (#4111)** — help / version / license / completions / doctor / easter egg など起動直後に完結する処理を通常の CLI dispatcher から分け、dispatch 境界をテスト・レビューしやすくしました。
